Baccus-2/5
I've added some stuff to tap test activity. I added a "try again"
button to restart the test. Once, it gets to five tries it shows a
button for the right hands test if you're on left hand test. And it
shows a button to send you to results if you completed the right hand
test. If you guys don't have any reservations about the additions I can
push the stuff I added.